requirements.txt
aiofiles==0.5.0
alembic==1.6.5
aniso8601==7.0.0
anyio==3.6.2
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.5.7
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.5
MarkupSafe==2.1.1
mangum==0.17.0
orjson==3.8.3
passlib==1.7.4
promise==2.3
psycopg2==2-2.9.5
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
rsa==4.9
Rx==1.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
toml==0.10.2
typing-extensions==4.4.0
ujson==5.6.0
urllib3==1.26.6
uvicorn==0.20.0
watchfiles==0.18.1
watchgod==0.7
websockets==10.4